# Implement string pooling in scanner/parser