Motivation:

> BTW, [this benchmark](https://github.com/sebastienros/esprima-dotnet/pull/313#issuecomment-1215379293) of yours showing that `string` and `Identifier` are sitting on the top of the list gave me an idea how we might push this even further:
> 
> Let's suppose we have the following input:
> 
> ```js
> let abc;
> abc = 0;
> ```
> 
> We declare something, then later we reference that using its name. Pretty common stuff. When we parse this, we end up with an AST containing two identical `Identifier`s (well, not completely identical because of metadata like location, range, etc.) However, the `Name` property of these identifiers will be exactly the same strings. I mean they'll contain the same characters. They won't currently be the same _instance_ (apart from some special cases covered by `ParserExtensions.Slice`) though. But they could be and this is where I see a room for improvement.
> 
> It looks possible to introduce some kind of a string pooling by means of which we could make sure that `Scanner` won't allocate the same string multiple times. We may borrow some ideas from [here](https://docs.microsoft.com/en-us/windows/communitytoolkit/high-performance/stringpool) as it looks to do exactly what I'm talking about.

First results:

## Esprima.Benchmark.FileParsingBenchmark

| **Diff**|Method|FileName|Mean|Error|Gen 0|Gen 1|Gen 2|Allocated|
|------- |-------|-------|-------:|-------|-------:|-------:|-------:|-------:|
| Old |ParseProgram|angular-1.2.5|21.162 ms|0.4196 ms|1000.0000|468.7500|-|5,744 KB|
| **New** |	|	| **20.491 ms (-3%)** | **1.2497 ms** | **937.5000 (-6%)** | **437.5000 (-7%)** | **-** | **5,553 KB (-3%)** |
| Old |ParseProgram|backbone-1.1.0|2.609 ms|0.4331 ms|160.1563|70.3125|-|888 KB|
| **New** |	|	| **2.706 ms (+4%)** | **0.0774 ms** | **148.4375 (-7%)** | **66.4063 (-6%)** | **-** | **867 KB (-2%)** |
| Old |ParseProgram|jquery-1.9.1|15.677 ms|1.6023 ms|812.5000|375.0000|-|4,991 KB|
| **New** |	|	| **16.503 ms (+5%)** | **2.2575 ms** | **812.5000 (0%)** | **375.0000 (0%)** | **-** | **4,854 KB (-3%)** |
| Old |ParseProgram|jquery.mobile-1.4.2|27.572 ms|29.7640 ms|1250.0000|625.0000|93.7500|7,677 KB|
| **New** |	|	| **25.710 ms (-7%)** | **5.6449 ms** | **1218.7500 (-3%)** | **593.7500 (-5%)** | **31.2500 (-67%)** | **7,349 KB (-4%)** |
| Old |ParseProgram|mootools-1.4.5|12.870 ms|1.5982 ms|703.1250|312.5000|-|4,028 KB|
| **New** |	|	| **13.191 ms (+2%)** | **0.9527 ms** | **671.8750 (-4%)** | **312.5000 (0%)** | **-** | **3,936 KB (-2%)** |
| Old |ParseProgram|underscore-1.5.2|2.197 ms|0.1964 ms|140.6250|62.5000|-|748 KB|
| **New** |	|	| **2.251 ms (+2%)** | **0.1881 ms** | **140.6250 (0%)** | **62.5000 (0%)** | **-** | **739 KB (-1%)** |
| Old |ParseProgram|yui-3.12.0|11.652 ms|0.8847 ms|656.2500|281.2500|-|3,789 KB|
| **New** |	|	| **11.899 ms (+2%)** | **0.6330 ms** | **609.3750 (-7%)** | **296.8750 (+6%)** | **-** | **3,620 KB (-4%)** |

These results are not as good as I expected. But there are some parameters to tweak (string pool initial buffer size, buffer growth rate, etc) and the hash algorithm can be optimized further as well. We'll see if it can get any better.