@sebastienros 2022-08-17
The TextSpan strategy might have a disadvantage compared to interned strings in the way that the source of the script will always be referenced by the final AST once the parsing is done. In terms of memory there is no allocations, but it might hide the fact that a bigger chunk of memory can't be freed. Micro-benchmarks won't show that. Might be interesting to have a column with memory delta before/after the method and a GC to show that.

One could argue that it's using more memory (after parsing), but it's still faster and still allocating less.