@sebastienros 2022-08-17
FYI, Ben Adams made an interning library https://github.com/benaadams/Ben.StringIntern