@adams85 2022-08-18
I've found some low hanging fruits around `Scanner.ScanPunctuator`:

## Esprima.Benchmark.FileParsingBenchmark

| **Diff**|Method|FileName|Mean|Error|Gen 0|Gen 1|Gen 2|Allocated|
|------- |-------|-------|-------:|-------|-------:|-------:|-------:|-------:|
| Old |ParseProgram|angular-1.2.5|21.056 ms|2.2627 ms|1000.0000|468.7500|-|5,744 KB|
| **New** |	|	| **19.723 ms (-6%)** | **1.2730 ms** | **718.7500 (-28%)** | **343.7500 (-27%)** | **-** | **4,391 KB (-24%)** |
| Old |ParseProgram|backbone-1.1.0|2.595 ms|0.1273 ms|164.0625|78.1250|-|888 KB|
| **New** |	|	| **2.759 ms (+6%)** | **0.4853 ms** | **140.6250 (-14%)** | **42.9688 (-45%)** | **-** | **679 KB (-24%)** |
| Old |ParseProgram|jquery-1.9.1|15.487 ms|0.4387 ms|875.0000|437.5000|-|4,991 KB|
| **New** |	|	| **15.798 ms (+2%)** | **0.8879 ms** | **625.0000 (-29%)** | **312.5000 (-29%)** | **-** | **3,770 KB (-24%)** |
| Old |ParseProgram|jquery.mobile-1.4.2|27.276 ms|28.5918 ms|1312.5000|656.2500|187.5000|7,677 KB|
| **New** |	|	| **24.657 ms (-10%)** | **0.5061 ms** | **1000.0000 (-24%)** | **500.0000 (-24%)** | **31.2500 (-83%)** | **5,800 KB (-24%)** |
| Old |ParseProgram|mootools-1.4.5|12.754 ms|1.5094 ms|703.1250|312.5000|-|4,028 KB|
| **New** |	|	| **12.707 ms (0%)** | **4.7469 ms** | **500.0000 (-29%)** | **250.0000 (-20%)** | **-** | **3,095 KB (-23%)** |
| Old |ParseProgram|underscore-1.5.2|2.370 ms|0.0625 ms|132.8125|54.6875|-|748 KB|
| **New** |	|	| **2.290 ms (-3%)** | **0.2174 ms** | **121.0938 (-9%)** | **39.0625 (-29%)** | **-** | **574 KB (-23%)** |
| Old |ParseProgram|yui-3.12.0|11.833 ms|2.5972 ms|656.2500|296.8750|-|3,789 KB|
| **New** |	|	| **11.487 ms (-3%)** | **1.0623 ms** | **468.7500 (-29%)** | **234.3750 (-21%)** | **-** | **2,853 KB (-25%)** |

It looks now we're getting somewhere. There are still some knobs to tweak but I expect that memory allocation can't significantly be reduced further. I'm more optimistic about execution time because of the more advanced ROS hashing methods.

Also, the latest commit includes improvements regarding https://github.com/sebastienros/esprima-dotnet/pull/314#discussion_r948267245