@lahma 2022-08-18
> > FYI, Ben Adams made an interning library https://github.com/benaadams/Ben.StringIntern
> 
> Thanks for the tip. I've checked it out and found that the implementation is based on BCL's `HashSet<T>` just like ours but it's a more generic one than that and does _much_ more than what we need here. Which means much higher complexity, so I doubt we'd get better results from switching to this lib.
> 
> One thing we may consider borrowing from there is [the Marvin alg](https://github.com/benaadams/Ben.StringIntern/blob/716cfbba5338ce070aafbeafbdd549395649f575/src/Helpers/Marvin.cs#L12) for hashing `ReadOnlySpan<char>` (but this would involve pulling in `System.Runtime.CompilerServices.Unsafe`). However, [this alg](https://github.com/CommunityToolkit/WindowsCommunityToolkit/blob/winui-7.1.2/CommunityToolkit.HighPerformance/Helpers/Internals/SpanHelper.Hash.cs#L87) looks even more powerful. This will definitely need some benchmarking.

Those are cool implementations. I actually tried the WindowsCommunityToolkit version briefly on Jint side a while back, can't remember specifics though 😄 

> > The TextSpan strategy might have a disadvantage compared to interned strings in the way that **the source of the script will always be referenced by the final AST once the parsing is done**.
> 
> The fact that the entire input would be kept around (which could be take up megabytes) until even a single node which has a reference to it is alive looks like a showstopper to me.
> 
> I see two major problems with the TextSpan approach:
> 
> 1. Our parser doesn't collect whitespace trivia like e.g. Roslyn, so in the case of larger files we'd keep around a lot of unreferenced whitespace unnecessary. (And this goes for tokens and comments when collection of them is turned off in the options.)
> 2. Even if our parser collected whitespace trivia, the concept would begin to fall apart as soon as we consider AST rewriting/modification. When a large part of the tree is rewritten, we can easily end up with the unfortunate situation outlined above.
> 
> @lahma What do you think about these arguments? Am I missing here something maybe?

This is *a very good* point. I'm starting to lean towards string interning + recently used pooling approach that you have here. Maybe we'll ditch the StringSegment approach unless something really interesting comes along later.

Is ScanPunctuator work something that could be brough to main separately to focus on the differences the pooling in this PR brings, would it make sense? Some of the diffs look familiar like char to string lookup, I probably also experimented with those on my branch.