@adams85 2022-08-18
> Is ScanPunctuator work something that could be brough to main separately to focus on the differences the pooling in this PR brings, would it make sense? 

It would make perfect sense. Will do it soon.