@adams85 2022-08-18
So, it seems this is what (a plain and simple) string pooling can bring to the table:

## Esprima.Benchmark.FileParsingBenchmark

| **Diff**|Method|FileName|Mean|Error|Gen 0|Gen 1|Gen 2|Allocated|
|------- |-------|-------|-------:|-------|-------:|-------:|-------:|-------:|
| Old |ParseProgram|angular-1.2.5|19.870 ms|0.9480 ms|750.0000|375.0000|-|4,718 KB|
| **New** |	|	| **19.869 ms (0%)** | **1.5634 ms** | **750.0000 (0%)** | **343.7500 (-8%)** | **-** | **4,392 KB (-7%)** |
| Old |ParseProgram|backbone-1.1.0|2.730 ms|0.1545 ms|140.6250|58.5938|-|721 KB|
| **New** |	|	| **2.848 ms (+4%)** | **0.5838 ms** | **140.6250 (0%)** | **46.8750 (-20%)** | **-** | **679 KB (-6%)** |
| Old |ParseProgram|jquery-1.9.1|15.617 ms|3.7405 ms|656.2500|328.1250|-|4,079 KB|
| **New** |	|	| **15.717 ms (+1%)** | **3.2395 ms** | **593.7500 (-10%)** | **281.2500 (-14%)** | **-** | **3,770 KB (-8%)** |
| Old |ParseProgram|jquery.mobile-1.4.2|24.085 ms|0.8815 ms|1031.2500|500.0000|-|6,304 KB|
| **New** |	|	| **24.907 ms (+3%)** | **3.6787 ms** | **1000.0000 (-3%)** | **468.7500 (-6%)** | **31.2500** | **5,800 KB (-8%)** |
| Old |ParseProgram|mootools-1.4.5|12.483 ms|1.4284 ms|546.8750|265.6250|-|3,290 KB|
| **New** |	|	| **12.588 ms (+1%)** | **1.0525 ms** | **500.0000 (-9%)** | **250.0000 (-6%)** | **-** | **3,095 KB (-6%)** |
| Old |ParseProgram|underscore-1.5.2|2.261 ms|0.0589 ms|113.2813|42.9688|-|603 KB|
| **New** |	|	| **2.331 ms (+3%)** | **0.0859 ms** | **117.1875 (+3%)** | **39.0625 (-9%)** | **-** | **574 KB (-5%)** |
| Old |ParseProgram|yui-3.12.0|11.954 ms|2.6612 ms|500.0000|250.0000|-|3,079 KB|
| **New** |	|	| **11.726 ms (-2%)** | **0.0871 ms** | **484.3750 (-3%)** | **234.3750 (-6%)** | **-** | **2,900 KB (-6%)** |

At least, I couldn't improve it further. I tested both Marvin and Djb algorithms for ROS hashing but it didn't make any difference. Probably because the most frequent pooled strings tend to be short.

As to @sebastienros's remark in the other PR:

> Note that parlot falls in this category at the moment. It makes me think that some allocations benchmarks are flawed since we don't take the preallocated memory into account when measuring an iteration.

> I haven't looked at the lifetime scope of the pool either in the intern pr. It matters in concurrent systems where the same pool might be in locking contention.

This implementation doesn't use pre-allocated memory. Each `Scanner` instance [starts with an own empty pool](https://github.com/sebastienros/esprima-dotnet/blob/26f3ce9a5bcc3e9232652b3c1075bea81bf49414/src/Esprima/Scanner.cs#L51) and [gets rid of that as soon as EOF is reached](https://github.com/sebastienros/esprima-dotnet/blob/26f3ce9a5bcc3e9232652b3c1075bea81bf49414/src/Esprima/Scanner.cs#L2433) (so the buffer can be GC'd even if the `Scanner` instance is kept alive.)

Because of this, we have no concurrency worries either. (A shared pool, however, is not something I'd consider because that looks pretty tricky and it's pretty dubious that we would have a clear gain from that in the end.)

Now it's up to you whether to merge this or not.